+++ hc/front/tests/test_send_test_notification.py
@@ -2,7 +2,7 @@ import json
 from unittest.mock import patch
 
 from django.core import mail
-from hc.api.models import Channel
+from hc.api.models import Channel, Notification
 from hc.test import BaseTestCase
 
 
@@ -30,6 +30,34 @@ class SendTestNotificationTestCase(BaseTestCase):
         self.assertEqual(email.to[0], "alice@example.org")
         self.assertTrue("X-Bounce-Url" in email.extra_headers)
         self.assertTrue("List-Unsubscribe" in email.extra_headers)
+
+        # It should create a notification
+        n = Notification.objects.get()
+        self.assertEqual(n.channel, self.channel)
+        self.assertEqual(n.error, "")
+
+    def test_it_clears_channel_last_error(self):
+        self.channel.last_error = "Something went wrong"
+        self.channel.save()
+
+        self.client.login(username="alice@example.org", password="password")
+        self.client.post(self.url, {})
+
+        self.channel.refresh_from_db()
+        self.assertEqual(self.channel.last_error, "")
+
+    def test_it_sets_channel_last_error(self):
+        self.channel.email_verified = False
+        self.channel.save()
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, {}, follow=True)
+
+        self.assertContains(r, "Could not send a test notification")
+        self.assertContains(r, "Email not verified")
+
+        self.channel.refresh_from_db()
+        self.assertEqual(self.channel.last_error, "Email not verified")
 
     @patch("hc.api.transports.requests.request")
     def test_it_handles_webhooks_with_no_down_url(self, mock_get):