+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 ### Bug Fixes
 - Removing Pager Team integration, project appears to be discontinued
+- Sending a test notification updates Channel.last_error (#391)
 
 ## v1.15.0 - 2020-06-04